Split JPA implementations.
